requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.11
gunicorn==20.1.0
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
requests==2.27.1
djangorestframework==3.13.1
Pillow==9.1.0
pandas==1.4.2
numpy==1.23.0
tensorflow==2.9.1
opencv-python==4.6.0.66
matplotlib==3.5.2
pygame==2.1.2
